requirements.txt
asgiref==3.9.1
attrs==25.3.0
babelfish==0.6.1
beautifulsoup4==4.13.4
certifi==2025.6.15
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
click-option-group==0.5.7
colorama==0.4.6
coverage==7.10.1
decorator==5.2.1
defusedxml==0.7.1
distlib==0.4.0
dj-database-url==2.3.0
Django==5.2.5
django-cors-headers==4.3.1
django-debug-toolbar==4.4.2
django-filter==24.3
django-taggit==6.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dogpile.cache==1.4.0
drf-spectacular==0.28.0
enzyme==0.5.2
factory_boy==3.3.3
Faker==37.4.2
filelock==3.18.0
flake8==7.3.0
getjump==2.7.3
guessit==3.8.0
gunicorn==21.2.0
identify==2.6.12
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
isort==6.0.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
knowit==0.5.11
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
nodeenv==1.9.1
packaging==25.0
pillow==10.4.0
platformdirs==4.3.8
pluggy==1.6.0
pre-commit==3.7.1
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pymediainfo==7.0.1
pypresence==4.3.0
pysubs2==1.8.0
pytest==8.4.1
pytest-cov==5.0.0
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
rebulk==3.2.0
referencing==0.36.2
requests==2.32.4
rich==13.9.4
rpds-py==0.26.0
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
srt==3.5.3
stevedore==5.5.0
subliminal==2.3.2
tomlkit==0.13.3
trakit==0.2.5
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
virtualenv==20.32.0
whitenoise==6.9.0
